WIP - propagating hierarchy changes in TimeSeries
