--- darts/reconciliation/__init__.py
@@ -1,4 +1,0 @@
-"""
-Posthoc Forecast Reconciliation
--------------------------------
-"""
